set up the authentication in correct files
