+++ components/Welcome.tsx
@@ -24,7 +24,7 @@ export default function Welcome() {
         </h1>
       </div>
       <Link href="/map" className="mt-16 mb-60 md:mt-24 lg:mt-32">
-        <button className="px-14 py-3 bg-purple font-semibold text-white rounded-full text-lg md:text-2xl md:py-4 md:px-28 lg:text-4xl lg:py-6 lg:px-40">
+        <button className="px-14 py-3 bg-purple font-semibold text-white rounded-full text-lg md:text-2xl md:py-4 md:px-28 lg:text-4xl lg:py-6 lg:px-40 active:bg-[#34287D] active:lg:py-5 active:lg:px-36 transition-all">
           Get Started
         </button>
       </Link>
