+++ styles/app.css
@@ -4,7 +4,7 @@
 
 body {
   margin: 0;
-  background: #FFF;
+  background: #f2d93d;
   display: flex;
   flex-direction: column;
   font-family: "Lato", sans-serif;
